# openapi.yaml
...
openapi: 3.0.3
info:
  title: Provenara
  version: 0.1.0
  description: >-
    Art-finance trust and fractional-ownership control plane for provenance, custody, screening, and share registers.
  x-product-slug: ai-fin-provenara
  x-source: ai-in-financial/deloitte-lu-art-finance-conf-presentation-2018/
servers:
  - url: https://api.example.com
...
paths:
  /v1/artworks:
    get:
      operationId: listArtworks
      tags: [Artworks]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeArtwork"
    post:
      operationId: createArtwork
      tags: [Artworks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeArtwork"
...
  /v1/artworks/{artworkId}/valuations:
    parameters:
      - $ref: "#/components/parameters/ArtworkId"
    post:
      operationId: addValuation
      tags: [Valuations]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValuationMarkCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeValuationMark"
...
  /v1/artworks/{artworkId}/anchors:
    parameters:
      - $ref: "#/components/parameters/ArtworkId"
    post:
      operationId: publishAnchor
      tags: [Audits]
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Anchor receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeAnchorReceipt"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    ArtworkId:
      name: artworkId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
...
    Artwork:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        artist: { type: string }
        year: { type: integer }
    ArtworkCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        artist: { type: string }
        year: { type: integer }
...
    ValuationMark:
      type: object
      required: [id, artworkId, amount, currency, valuedOn]
      properties:
        id: { type: string }
        artworkId: { type: string }
        amount: { type: number }
        currency: { type: string }
        valuedOn: { type: string, format: date }
        expiresOn: { type: string, format: date }
        method: { type: string }
        valuer: { type: string }
    ValuationMarkCreate:
      type: object
      required: [amount, currency, valuedOn, method, valuer]
      properties:
        amount: { type: number }
        currency: { type: string }
        valuedOn: { type: string, format: date }
        expiresOn: { type: string, format: date }
        method: { type: string }
        valuer: { type: string }
...
    AnchorReceipt:
      type: object
      required: [id, artworkId, contentHash]
      properties:
        id: { type: string }
        artworkId: { type: string }
        contentHash: { type: string }
        anchoredAt: { type: string, format: date-time }
...
    DataEnvelopeArtwork:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Artwork"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeArtwork:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Artwork"
            nextCursor: { type: string }
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeValuationMark:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ValuationMark"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeAnchorReceipt:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AnchorReceipt"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...