# platform/api-server/src/domains/artworks/handlers/artwork.handlers.ts
/**
 * Handlers for Artwork
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: artworks
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ArtworksDomainModule } from "../dependencies/artworks-ddd.dependencies.js";
import type { CreateArtworkInput, GetArtworkInput, ListArtworksInput } from "@provenara/services/artworks";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/artworks
 * Operation: listArtworks
 */
export async function listArtworks(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], readiness?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ArtworksDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListArtworksInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.artworks.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/artworks
 * Operation: createArtwork
 */
export async function createArtwork(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ArtworksDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CreateArtworkInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.artworks.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/artworks/{artworkId}
 * Operation: getArtwork
 */
export async function getArtwork(
  request: FastifyRequest<{ Params: { artworkId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ArtworksDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetArtworkInput = {
    artworkId: request.params.artworkId,
  };
  const result = await deps.useCases.artworks.get.execute(input);
  return reply.code(200).send(result);
}