# platform/api-server/src/domains/audits/handlers/anchor.handlers.ts
/**
 * Handlers for Anchor
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: audits
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { AuditsDomainModule } from "../dependencies/audits-ddd.dependencies.js";
import type { ListAnchorsInput, PublishAnchorInput } from "@provenara/services/audits";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/artworks/{artworkId}/anchors
 * Operation: listAnchors
 */
export async function listAnchors(
  request: FastifyRequest<{ Params: { artworkId: string }; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: AuditsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListAnchorsInput = {
    artworkId: request.params.artworkId,
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.anchors.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/artworks/{artworkId}/anchors
 * Operation: publishAnchor
 */
export async function publishAnchor(
  request: FastifyRequest<{ Params: { artworkId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: AuditsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: PublishAnchorInput = {
    artworkId: request.params.artworkId,
  };
  const result = await deps.useCases.anchors.get.execute(input);
  return reply.code(201).send(result);
}