# platform/api-server/src/domains/valuations/handlers/valuation.handlers.ts
/**
 * Handlers for Valuation
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: valuations
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ValuationsDomainModule } from "../dependencies/valuations-ddd.dependencies.js";
import type { AddValuationInput, ListValuationsInput } from "@provenara/services/valuations";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/artworks/{artworkId}/valuations
 * Operation: listValuations
 */
export async function listValuations(
  request: FastifyRequest<{ Params: { artworkId: string }; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ValuationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListValuationsInput = {
    artworkId: request.params.artworkId,
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.valuations.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/artworks/{artworkId}/valuations
 * Operation: addValuation
 */
export async function addValuation(
  request: FastifyRequest<{ Params: { artworkId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ValuationsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: AddValuationInput = {
    artworkId: request.params.artworkId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.valuations.create.execute(input);
  return reply.code(201).send(result);
}